Calculate orientation quaternion from accelerometer/magnetometer
This commit switches the firmware to returning an orientation
quaternion based soley on the accelerometer and magnetometer data.
Kalman filtering will be added later.

The orientation Processing sample was modified to expect this
4 x float quaternion instead of the 9 x float calibrated data I added
in the previous commit. It compares this quaternion to the one
calculated in the Processing code to verify that I didn't regress
anything as I ported this math to the embedded device.