Commit of quaternion and matrix dumping code
I just want to keep this code around incase I want to reference it in
the future. It is able to dump floats from various Kalman quaternions
and matrices right after the filter is reset to compare the values
between the PC and LPC1768.

NOTE: It does contain a few fixes which I will port back into master as
      well.