Port heading calculation to LPC1768
The compass heading calculation code is now running on the LPC1768 and
has been appended to the end of the list of measurements returned to
the PC one hundred times/second.

I also added code to the orientation Processing sample to compare the
heading calculated on the PC to the value calculated by the LPC1768 and
an error is displayed to the console if they differ.