Add simple drive sample
This sample instructs the Arlo platform to drive ahead at a speed of
100 encoder ticks per second and verifies that the overall distance
travelled (in encoder ticks) has increased as expected. This test is
run for 2 seconds and then the robot is stopped.

The bulk of the functionality is provided by the newly added
libs/MotorController library which knows how to talk to the Parallax
DHB-10 Dual Motor Controller on the Arlo.

I have also added the bash scripts that I use for remote debug/
programming of the LPC1768 on the robot.