Re-calibrate heading sensor after mounting on robot
The magnetometer calibration data had to be updated to account for hard
iron on the robot. I also adjusted the compass.mountingCorrection to
account for the fact that I mounted the project box 180 degrees from
the original orientation. The USB connector now points towards the
front of the robot and not the back.