Add DebugSerial library/class
This class can be used to extend the functionality of the MRI debug
monitor. In particular, it allows for the sending of debug output to
GDB in an async fashion without blocking the CPU. Without this class
output is sent to the GDB console using printf() which ends up
halting into the MRI debug monitor to start a dialog with GDB and it
remains halted until all of the text has been sent to GDB. When a
DebugSerial object is constructed, it will find the UART used by MRI
and hijack its ISR. The outputStringToGdb() method can be called to
utilize DMA to asynchronously send the O (output) packet to GDB. The
ISR shim will detect GDB sending back acks for these O packets and
ignore them. If GDB sends CTRL+C to break into the debuggee, then this
ISR shim will force a halt into the MRI debug monitor.