Initial Robo-Magellan robot firmware
This code doesn't really do much at this point. It does pull in the
Sparkfun9DoFSensorStick and MotorController classes that will be needed
by the robot for initial navigation tests. It also has the start of the
debugging framework that I plan to use.

It takes advantage of the DebugSerial class for sending messages to the
GDB console in an asynchronous fashion to minimize the impact on
performance.