Rotate robot inplace for PID modelling
The robot firmware now runs in a mode where it oscillates back and
forth 90 degrees around its original orientation. While doing this
oscillation it logs the time, current heading angle from the
orientation sensor, and the control output being sent to the motors.
Issuing the "set logging on" command to GDB before starting this test
will cause the logged data to be recorded in gdb.txt.  I then edit
this logged output to only include the comma separated data found
between these two lines:
    Starting oscillations between ##.## and ##.##...
        ...
    PID Calibration completed.
This results in many lines of data like the following:
    19,7.536273,100
    39,7.537899,100
    59,7.539880,100
    79,7.536492,100

The tools/pidCalc utility can be run on the data after editing to
automatically calculate the values required for a system model and
the resulting PI constants:
    ==Stats==
    slope1 = 1.507058
    slope2 = -1.549744
    CO1 = 100
    CO2 = -100
    Kp* = 0.015284
    Өp = 0.370527
    Tc = 1.111580
    Kc = 77.254082
    Ti = 2.593686

NOTE:
It is best to orient the robot so that its front points roughly north
when starting this calibration so that pidCalc doesn't fail around the
+180 degree/-180 degree discontinuity in the south.

I also updated the comments in the __mriPlatform_*DebuggerHook()
routines to make it clearer what they are doing.